LAST COMMIT BEFORE LEAVING
git-svn-id: file:///home/roberto/Git/svn@10745 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 0e3f96c01b91937dae13bf63f093791a3aefae93

+++ trunk/xmipp/libraries/data/micrograph.h
@@ -598,7 +598,7 @@ public:
 	void computeGamma();
 
 	/// Compute alphas
-	void computeAngles(int &ualpha, int &talpha, int &ogamma);
+	void computeAngles(double &ualpha, double &talpha, double &ogamma);
 public:
     // For tilted-untilted correspondance
     Matrix2D<double>    Au;     // Untilted "positions"
